requirements.azurefunction.txt
azure-functions
nest_asyncio
